Fix private tangle permissions (#1390)
* - fix dir permissions when running private tangle

* - Build latest image when bootstrapping private tangle

* - Create all dirs beforehand to do chown

* - Add state dir

* - Delete all dirs on cleanup

- Create all dirs on bootstrap
- Warn if running without bootstrap

* - Add note to run cleanup with sudo

* - Cleaned up scripts and check for proper root permissions to be able to set dir permissions

* - Removed wait_for_inx docker hubs and using 0.1 instead of latest for indexer and mqtt

* - Create coo snapshot dir from script

* - Create all privatedb folders

* - Do not require sudo and do chown on macOS

* - Re-enabled build

* - Added run_rocker.sh skript to run the normal docker setup

- calling pull with image names to avoid error warnings in the logs

* - Use build arg in run_docker.sh

* - Added build flag to bootstrap.sh

+++ private_tangle/cleanup.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
 
-docker-compose run cleanup
+if [[ "$OSTYPE" != "darwin"* && "$EUID" -ne 0 ]]; then
+  echo "Please run as root or with sudo"
+  exit
+fi
+
 docker-compose down --remove-orphans
+
+rm -Rf privatedb
+rm -Rf snapshots
